requirements.txt
matplotlib==2.1.0
numpy==1.16.1
Pillow==4.2.1
scipy==1.2.0
torch==0.4.1
torchvision==0.2.1
tqdm==4.28.1
